% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupProjFunctions.R
\name{age_gender_plots}
\alias{age_gender_plots}
\title{Age Gender Plot by Country/Region}
\usage{
age_gender_plots(entity, dataframe)
}
\arguments{
\item{entity}{enter the country or region you want as a string}

\item{entityDF}{the data frame you want to use}
}
\value{
a plot that has the mental health prevalence rates by
age and gender country or region
}
\description{
Function that returns all non-unique values in a vector
}
\examples{
age_gender_plots("United States", fullDF_region)
}
