% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupProjFunctions.R
\name{build_age_graphs}
\alias{build_age_graphs}
\title{Build an Age Graph}
\usage{
build_age_graphs(age_grp, df)
}
\arguments{
\item{age_grp}{age group to graph}
}
\value{
a line graph that shows the prevalence of depression across all regions for the specified age groups
}
\description{
Function that creates a line graph of the prevalence of depression for a certain age group in all regions
}
\examples{
age_grp = ten_to_fourteen
build_age_graphs(df$ten_to_fourteen, df)

}
