% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupProjFunctions.R
\name{country_region_plots}
\alias{country_region_plots}
\title{Region/Country Plot with Mental Health Disorder Prevalence}
\usage{
country_region_plots(entity, entityDF)
}
\arguments{
\item{entity}{enter the country or region you want as a string}

\item{entityDF}{the data frame you want to use}
}
\value{
a plot that has the mental health prevalence rates per country or region
}
\description{
Function that returns a plot of mental health prevalence rates
per country or region
}
\examples{
country_region_plots("United States", fullDF_region)

}
