% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupProjFunctions.R
\name{graph_for_region}
\alias{graph_for_region}
\title{Region Graphs by Age Group}
\usage{
graph_for_region(entity, df)
}
\arguments{
\item{entity}{name of entity}

\item{df}{name of dataframe}
}
\value{
line plot of prevalence for each age group
}
\description{
Function to create a plot that shows the prevalence of
each age group in a specific region over time
}
\examples{
entity = North America
df = northAmerica_pivot
graph_for_region("North America", northAmerica_pivot)

}
