% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupProjFunctions.R
\name{pivot_my_data}
\alias{pivot_my_data}
\title{Pivot my Data}
\usage{
pivot_my_data(entity, df)
}
\arguments{
\item{entity}{name of entity}
}
\value{
a pivoted dataframe with the Year in a column,
the age groups in a separate column, and the prevalence in a
column for one region where the region is the value passed in as entity
}
\description{
Function that pivots the data to put Year in a column,
age group in a column, and prevalence in a column
}
